python_scripts/requirements.txt
boto==2.44.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
crcmod==1.7
cryptography==1.7.1
enum34==1.1.6
futures==3.2.0
google-api-core==1.10.0
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-automl==0.2.0
google-cloud-language==1.2.0
google-cloud-speech==1.0.0
google-compute-engine==2.8.13
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
ipaddress==1.0.17
keyring==10.1
keyrings.alt==1.3
numpy==1.16.3
oauthlib==3.0.1
pandas==0.24.2
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycrypto==2.6.1
pygobject==3.22.0
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2019.1
pyxdg==0.25
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
SecretStorage==2.3.1
six==1.10.0
tweepy==3.7.0
urllib3==1.24.3
